Giphy function (#11)
* feat(giphy): adding a giphy api capability

This will be used by other functions in the future, like cheering on karma earns, but for now this
is just laying the groundwork. Next commit will also include efforts to create a basic giphy based
tool, /cheer.

* fix(giphy): /cheer now works using the giphy function

The other changes were all structural differences setup to enable better separation of data and
usage of functions and structures going forward.